[IAE-3167] Mark sidenav links as buttons for screenreaders
Screenreaders are not interpreting our <a> tags as selectable,
possibly because we do not use `href=`. Adding `role=button`
will allow screenreaders to treat them as buttons.

This is a temporary fix; we need to decide whether to keep using
<a> tags or switch to using <button> which would be more semantic,
but we would need to rewrite our css rules for buttons. If we
decide to keep using <a> tags, then we should implement an event
handler for spacebar.

More information here:
https://formidable.com/blog/2014/05/08/anchors-buttons-and-accessibility/